tests: fix utils placement
