+++ tests/func/metrics/plots/test_modify.py
@@ -4,7 +4,7 @@ from dvc.dvcfile import PIPELINE_LOCK
 from dvc.repo.plots import PropsNotFoundError
 from dvc.repo.plots.template import TemplateNotFoundError
 from dvc.utils import relpath
-from tests.func.metrics import _write_json
+from tests.func.metrics.utils import _write_json
 
 
 def test_plots_modify_existing_template(
